ml/requirements.txt
ultralytics==8.3.23
langchain==0.2.16
langchain-ollama==0.1.3
pillow